+++ CHANGES.md
@@ -1,8 +1,9 @@
 ## 0.0.2
 - add `Term.destruct2`
 - add `Pattern.var_any`
-- comment fixes
 - add `Zipper` module
+- add `Susbt.Unification.unfold`
+- comment fixes
 
 ## 0.0.1
 - First release of `term-indexing`, a library to index and rewrite sets of first-order terms.